Application/ml-model/requirements.txt
flask==2.0.1
requests==2.26.0
Werkzeug==2.0.1
dnspython
requests
pandas
scikit-learn